Some ruby gem updates and cleanup
