+++ packages/buildessential.rb
@@ -171,8 +171,6 @@ class Buildessential < Package
   depends_on 'ruby_rubocop'
   # Add ruby_concurrent_ruby
   depends_on 'ruby_concurrent_ruby'
-  # For use in ruby prompts.
-  depends_on 'ruby_highline'
   # For crew debugging.
   depends_on 'ruby_pry_byebug' if CREW_IN_CONTAINER
