+++ tools/packages.yaml
@@ -7871,6 +7871,11 @@ url: https://github.com/markdownlint/markdownlint/tags
 activity: low
 ---
 kind: url
+name: ruby_pry
+url: https://github.com/pry/pry/tags
+activity: low
+---
+kind: url
 name: ruby_pry_byebug
 url: https://github.com/deivid-rodriguez/pry-byebug/releases
 activity: low
